Split test_main into per-command files
